requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
SQLAlchemy==2.0.44
alembic==1.17.0
pydantic==2.12.2
python-jose==3.5.0
python-dotenv==1.1.1
pydantic-settings==2.11.0
asyncpg==0.30.0
email-validator==2.3.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.0.1